# swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Tienda y Pasarela de Pago (TPP)
  description: Esta es la documentación para la API del Microservicio de Tienda y
    Pasarela de Pago (TPP) del proyecto OverSounds.
  version: 2.0.0
...
servers:
- url: http://localhost
...
paths:
...
  /purchase:
    post:
      tags:
      - purchases
      summary: Set a product as purchased by the user.
      description: Set a product as purchased by the user.
      operationId: set_purchase
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Purchase"
      responses:
        "200":
          description: Successful operation
        "422":
          description: Validation exception
        "500":
          description: Generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - oversound_auth:
        - write:purchases
      x-openapi-router-controller: swagger_server.controllers.purchases_controller
components:
  schemas:
...
    Purchase:
      required:
      - paymentMethodId
      - purchaseDate
      - purchasePrice
      type: object
      properties:
        purchasePrice:
          type: number
          description: Price of the Song
          format: float
          example: 10.99
        purchaseDate:
          type: string
          description: Date of the purchase
          format: date-time
          example: 2010-01-15T10:30:00Z
        paymentMethodId:
          type: integer
          description: id of the payment method
          example: 0
        songIds:
          type: array
          items:
            type: integer
          description: List of song IDs to purchase
          example: [1, 2, 3]
        albumIds:
          type: array
          items:
            type: integer
          description: List of album IDs to purchase
          example: [1, 2]
        merchIds:
          type: array
          items:
            type: integer
          description: List of merch IDs to purchase
          example: [1]
      description: A purchase made by an user.
      example:
        purchasePrice: 10.99
        purchaseDate: 2010-01-15T10:30:00Z
        paymentMethodId: 0
        songIds: [1, 2, 3]
        albumIds: [1, 2]
        merchIds: [1]
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
  securitySchemes:
    oversound_auth:
      type: apiKey
      name: token
      in: cookie
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_oversounds_auth